move into wiki path
