+++ packages/design/components/Form/style/index.less
@@ -9,7 +9,7 @@
     &-label {
       display: inline-block;
       font-weight: 400;
-      font-size: 18px;
+      font-size: 1.125rem;
       line-height: 25px;
       color: #9f9b9b;
