+++ packages/design/components/Input/style/index.less
@@ -45,7 +45,7 @@
   color: @gray-100;
   padding: 0;
   width: 100%;
-  height: 100%;
+  height: 1rem;
 
   &::placeholder {
     color: @gray-10;
