+++ packages/design/components/Radio/style/index.less
@@ -18,4 +18,9 @@
     user-select: none;
     font-size: 1rem;
   }
+
+  &-group {
+    display: inline-flex;
+    gap: 0.5rem;
+  }
 }
