+++ packages/design/components/Select/style/index.less
@@ -4,7 +4,7 @@
   display: inline-flex;
   box-sizing: border-box;
   position: relative;
-  padding: 8px 12px;
+  padding: 0.5rem 1rem;
   border: 2px solid @gray-10;
   border-radius: 19px;
   user-select: none;
