+++ packages/website/src/components/Link/index.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren } from 'react';
 import React, { createContext, forwardRef, useContext, useTransition } from 'react';
-import type { LinkProps } from 'react-router-dom';
-import { useHref, useLinkClickHandler } from 'react-router-dom';
+import type { LinkProps, NavLinkProps } from 'react-router-dom';
+import { useHref, useLinkClickHandler, useLocation, useResolvedPath } from 'react-router-dom';
 
 interface ILinkContext {
   pending: boolean;
@@ -46,3 +46,46 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
 );
 
 export const useLinkContext = () => useContext(LinkContext);
+
+/**
+ * NavLink
+ * https://github.com/remix-run/react-router/blob/main/packages/react-router-dom/index.tsx#L470
+ */
+export const NavLink = React.forwardRef<HTMLAnchorElement, NavLinkProps>(function NavLinkWithRef(
+  {
+    'aria-current': ariaCurrentProp = 'page',
+    caseSensitive = false,
+    className: classNameProp = '',
+    end = false,
+    style: styleProp,
+    to,
+    children,
+    ...rest
+  },
+  ref,
+) {
+  const location = useLocation();
+  const path = useResolvedPath(to);
+  const { pending } = useLinkContext();
+
+  let locationPathname = location.pathname;
+  let toPathname = path.pathname;
+  if (!caseSensitive) {
+    locationPathname = locationPathname.toLowerCase();
+    toPathname = toPathname.toLowerCase();
+  }
+
+  const isActive =
+    locationPathname === toPathname ||
+    (!end &&
+      locationPathname.startsWith(toPathname) &&
+      locationPathname.charAt(toPathname.length) === '/');
+
+  const ariaCurrent = isActive ? ariaCurrentProp : undefined;
+
+  return (
+    <Link {...rest} aria-current={ariaCurrent} ref={ref} to={to}>
+      {typeof children === 'function' ? children({ isActive, isPending: pending }) : children}
+    </Link>
+  );
+});
