+++ packages/website/src/pages/index/[...slug].tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFound = () => {
+  return <>404</>;
+};
+
+export default NotFound;
