+++ packages/website/src/pages/index/subject/[id]/components/WikiLayout.tsx
@@ -2,35 +2,32 @@ import type { PropsWithChildren } from 'react';
 import React from 'react';
 
 import { Tab } from '@bangumi/design';
-import { useTransitionNavigate } from '@bangumi/website/hooks/use-navigate';
+import { NavLink } from '@bangumi/website/components/Link';
+import { WikiEditTabsItems } from '@bangumi/website/shared';
 
-import type { WikiEditTabs } from '../_shared';
-import { WikiEditTabsItems } from '../_shared';
 import style from './WikiLayout.module.less';
 
 type Props = PropsWithChildren<{
   id: number;
-  activeTab: WikiEditTabs;
   name: string;
 }>;
 
-const WikiLayout = ({ children, id, activeTab, name }: Props) => {
-  const [, navigate] = useTransitionNavigate();
-
+const WikiLayout = ({ children, id, name }: Props) => {
   return (
-    <div>
+    <>
       <div className={style.title}>
         正在编辑「<span className={style.subjectName}>{name}</span>」
       </div>
       <div className={style.subTitle}>WIKI STAFF ONLY</div>
-      <Tab
-        items={WikiEditTabsItems}
-        activeKey={activeTab}
-        type='borderless'
-        onChange={(_, value) => navigate(value.to(id))}
-      />
+      <Tab.Group type='borderless'>
+        {WikiEditTabsItems.map((item) => (
+          <NavLink to={item.to(id)} key={item.key}>
+            {({ isActive }) => <Tab.Item isActive={isActive}>{item.label}</Tab.Item>}
+          </NavLink>
+        ))}
+      </Tab.Group>
       <div className={style.body}>{children}</div>
-    </div>
+    </>
   );
 };
