+++ packages/website/src/pages/index/subject/[id]/wiki/common.module.less
@@ -32,11 +32,6 @@
   &Radio {
     padding-top: 9px;
     padding-bottom: 9px;
-
-    &Group {
-      display: flex;
-      gap: 10px;
-    }
   }
 
   &TextArea {
