+++ packages/website/src/pages/index/subject/[id]/wiki/history.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const WikiHistoryPage = () => {
+  return <>历史</>;
+};
+
+export default WikiHistoryPage;
